double and int can't be mixed together

+++ src/item.cpp
@@ -7448,7 +7448,7 @@ void item::process_temperature_rot( float insulation, const tripoint &pos,
             double env_temperature = 0;
             if( pos.z >= 0 ) {
                 w_point weather = wgen.get_weather( pos, time, seed );
-                env_temperature = std::max( weather.temperature + enviroment_mod + local_mod, -459 );
+                env_temperature = std::max( weather.temperature + enviroment_mod + local_mod, static_cast<double>(-459) );
             } else {
                 env_temperature = std::max( AVERAGE_ANNUAL_TEMPERATURE + enviroment_mod + local_mod, -459 );
             }
